Extract project layout helpers into `testing` subproject.
