+++ runtime/src/test/java/dev/ionfusion/fusion/RuntimeTest.java
@@ -4,8 +4,8 @@
 package dev.ionfusion.fusion;
 
 import static dev.ionfusion.fusion.FusionVoid.isVoid;
-import static dev.ionfusion.fusion.TestSetup.testDataFile;
-import static dev.ionfusion.fusion.TestSetup.testScriptDirectory;
+import static dev.ionfusion.testing.ProjectLayout.testDataFile;
+import static dev.ionfusion.testing.ProjectLayout.testScriptDirectory;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.allOf;
 import static org.hamcrest.Matchers.containsString;
